destiny-cli/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "destiny-cli"
version = "0.1.0"
description = "Command-line tools for Destiny Team Framework"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Destiny Team Framework", email = "team@destiny-framework.dev"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
    "psycopg2-binary>=2.9.0",
]

[project.scripts]
destiny = "destiny_cli.main:main"
